requirements.txt
fastapi
uvicorn
python-dotenv
jinja2
pydantic[email]
pytest
httpx
